+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### v1.58.2
+*2025-12-30*
+
+- Detect mobile/desktop version of site on page load to correctly add 8chanSS menu button
+- Fixes for Thread Watcher (You) mentions and highlights
+
+
 ### v1.58.1
 *2025-12-30*